reservation_system/requirements.txt
phonenumbers==8.12.22
Django==3.2
djangorestframework==3.12.2
drf-spectacular==0.13.2
python-dateutil==2.8.2
django-model-utils==4.2.0